Updated version number
On Android, removed any references to Activity or NavbarFragment from OnCreate and OnCreateView. It is only safe to use these in OnResume and later.
Additionally, we now check for the existence of any fragment before creating a new one.
This fixes crashing when Android unloads random fragments while the app is backgrounded.